requirements.txt
aniso8601==9.0.1
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
